% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccovid.R
\docType{data}
\name{bccovid}
\alias{bccovid}
\alias{bccovid_train}
\alias{bccovid_test}
\alias{bcpop}
\title{COVID 19 Cases in BC for the last year}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1143 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1655 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 685 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\source{
This data comes from the British Columbia CDC
\href{http://www.bccdc.ca/Health-Info-Site/Documents/BCCDC_COVID19_Dashboard_Case_Details.csv}{linelist}
via the \href{https://mountainmath.github.io/CanCovidData/}{\code{{CanCovidData}}} package.
}
\usage{
bccovid

bccovid_train

bccovid_test

bcpop
}
\description{
This data is the number of reported COVID-19 cases in British
Columbia between January 2020 and April 15, 2023. The values are.up-to-date
as of August 2023.
}
\details{
The full data (\code{bccovid}), contains the case count for BC.

The other data has been processed from this raw file. First, it is separated
out by Health Authority. Then it was
"smoothed" by applying a 7 day trailing sum. By this we mean that,

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cases_today  = sum(cases_7days_ago:cases_today)
}\if{html}{\out{</div>}}

The result is in the \code{cases+0} column. The other columns are "shifts" of this
value. \code{cases+14} is the response: it is the total number of cases observed
in the week ending 14 days after \code{date}. The other \code{cases} variables are
features: the total number of cases in the week ending on the \code{date}
(\code{cases+0}), 1 week earlier (\code{cases-7}) and 2 weeks earlier (\code{cases-14}).
\code{HA} is the name of the BC Health Authority. There are 5 different \code{HA}s,
and UBC is located in "Vancouver Coastal" along with Kitsilano and Downtown.
SFU, Burnaby, Richmond, and Surrey are all in "Fraser".

\code{bccovid_train} is the Training set for use on the homework.
It contains only the data prior to 1 August 2021.

\code{bccovid_test} is the Test set for use on the homework.
It contains 137 rows
of the data (between 1 August 2021 and 1 December 2021).

\code{bcpop} gives the number of people living in each BC Health Authority as of
the 2020 Census
}
\examples{
bccovid
bccovid_train
bccovid_test
bcpop
}
\keyword{datasets}
