% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost.R
\docType{data}
\name{goodboost}
\alias{goodboost}
\alias{badboost}
\title{Anonymous data for boosting exploration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 506 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 150 rows and 3 columns.
}
\usage{
goodboost

badboost
}
\description{
This is some made up data for exploring the performance of boosting.
The variables have been renamed and the class labels altered to remove
any historical associations with some other data that we try to avoid
using.

For both datasets, column names are \code{x}, \code{y}, and \code{class}.
}
\details{
\code{goodboost} is a case where stump-based boosting works well.

\code{badboost} is a case where stump-based boosting works poorly
}
\examples{
goodboost

badboost

}
\keyword{datasets}
