% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcycle.R
\docType{data}
\name{mcycle}
\alias{mcycle}
\alias{mcycle_train}
\alias{mcycle_test}
\title{Data from a Simulated Motorcycle Accident}
\format{
An object of class \code{data.frame} with 89 rows and 2 columns.

An object of class \code{data.frame} with 44 rows and 2 columns.
}
\source{
Silverman, B. W. (1985) Some aspects of the spline smoothing approach to
  non-parametric curve fitting.
  \emph{Journal of the Royal Statistical Society series B} \bold{47}, 1--52.
}
\usage{
mcycle_train

mcycle_test
}
\description{
A data frame giving a series of measurements of head acceleration
  in a simulated motorcycle accident, used to test crash helmets.
}
\details{
This data is originally available as \code{\link[MASS:mcycle]{MASS::mcycle}}.
}
\examples{
mcycle_train

mcycle_test

}
\references{
Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S-PLUS.} Fourth Edition. Springer.
}
\keyword{datasets}
