% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat406-package.R
\docType{package}
\name{Stat406-package}
\alias{Stat406}
\alias{Stat406-package}
\title{Stat406: Loads everything you need for Stat 406}
\description{
This package is a one-stop installation for all R packages and data used in UBC's Stat 406 course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ubc-stat.github.io/stat-406-rpackage/}
  \item \url{http://github.com/ubc-stat/stat-406-rpackage/}
  \item Report bugs at \url{https://github.com/UBC-STAT/stat-406-rpackage/issues/new}
}

}
\author{
\strong{Maintainer}: Daniel J. McDonald \email{daniel@stat.ubc.ca}

}
\keyword{internal}
