% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_latex_installation.R
\name{test_latex_installation}
\alias{test_latex_installation}
\title{Test for installed LaTeX distribution}
\usage{
test_latex_installation()
}
\description{
This function tries to compile a small pdf document using LaTeX.
If it fails, you'll see some instructions to make sure that LaTeX is
installed on your system.
}
