% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggly_function.R
\docType{data}
\name{wiggly}
\alias{wiggly}
\title{Data from \code{\link[=wiggly_function]{wiggly_function()}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 2 columns.
}
\usage{
wiggly
}
\description{
A small data set created using \code{\link[=wiggly_function]{wiggly_function()}}.
Used on Homework 4.
}
\keyword{datasets}
