% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggly_function.R
\name{wiggly_function}
\alias{wiggly_function}
\title{Wiggly function}
\usage{
wiggly_function(x)
}
\arguments{
\item{x}{Real numbers. Function is most useful for x in \verb{[0, 8]}}
}
\value{
Real numbers \deqn{\sin(x) + 1 / \sqrt(|x|) + 3}
}
\description{
Wiggly function
}
\examples{
curve(wiggly_function(x), 0, 8)
}
