% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggly_function.R
\docType{data}
\name{wiggly}
\alias{wiggly}
\title{Data from \code{\link[=wiggly_function]{wiggly_function()}}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 2 columns.
}
\usage{
wiggly
}
\description{
A small data set created using \code{\link[=wiggly_function]{wiggly_function()}}.
}
\examples{
wiggly
plot(wiggly, col = 2, pch = 16)
curve(wiggly_function(x), 0, 2 * pi, add = TRUE, col = 4, lwd = 2)
}
\keyword{datasets}
